refactor: extract playback engine from V3 player
